lr8 part1

+++ src/main/java/ua/com/kisit/course2026np/repository/FlightRepository.java
@@ -12,7 +12,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface FlightRepository extends JpaRepository<Flight, Integer> {
+public interface FlightRepository extends JpaRepository<Flight, Long> {
 
     // Пошук рейсу за унікальним номером
     Optional<Flight> findByFlightNumber(String flightNumber);
